Fix battery voltage reading for Inkplate 6Color
The battery voltage could not be read before, due to a change in the
port multiplexer chip used by the new inkplate 6color boards. This
change was not reflected by the base library, which therefore did not
enable the mosfet to activate the voltage line from the battery to allow
for measuring. This fix is only temporary until the corresponding code
is integrated into the real library

Ticket: https://github.com/SolderedElectronics/Inkplate-Arduino-library/issues/169